Create CGPA Calcualtor
The "CGPA Calculator with C++" is a simple yet effective command-line tool designed to help students calculate their Cumulative Grade Point Average (CGPA) based on the grades and credit hours of their courses. This project is implemented in C++ and provides an easy-to-use interface for inputting course details and obtaining the resulting CGPA.

Features
Course Data Input: Users can input the details of each course, including the course name, credit hours, and the corresponding grade (on a scale of 0 to 4).

CGPA Calculation: The program calculates the CGPA by taking into account the total grade points and credit hours of all entered courses.

Data Validation: The application ensures data integrity by validating the input for grade values, prompting the user to enter valid grades within the range of 0 to 4.